+++ jmetal/core/operator.py
@@ -16,6 +16,13 @@ class Operator(Generic[Source, Result]):
 class Mutation(Operator[Source, Source]):
     """ Class representing mutation operators """
 
+    def __init__(self, probability: float):
+        if probability > 1.0:
+            raise Exception("The probability is greater than one: " + str(probability))
+        elif probability < 0.0:
+            raise Exception("The probability is lower than zero: " + str(probability))
+
+        self.probability = probability
     def execute(self, source: Source) -> Source:
         pass
