+++ jmetal/core/problem.py
@@ -1,4 +1,5 @@
 from typing import Generic, TypeVar
+
 from jmetal.core.solution import BinarySolution, FloatSolution
 
 __author__ = "Antonio J. Nebro"
