--- jmetal/core/util/observer/test/__init__.py
@@ -1,2 +1,0 @@
-from jmetal.core.util.observer.observable import Observable
-from jmetal.core.util.observer.observer import Observer
