+++ jmetal/problem/multiobjective.py
@@ -1,4 +1,5 @@
 from math import sqrt, exp, pow, sin
+
 from jmetal.core.problem import FloatProblem
 from jmetal.core.solution import FloatSolution
