+++ jmetal/problem/test/viennet2Test.py
@@ -2,6 +2,7 @@ import unittest
 
 from jmetal.problem.multiobjective.viennet2 import Viennet2
 
+
 class TestMethods(unittest.TestCase):
 
     def setUp(self):
