+++ jmetal/runner/singleobjective/generationalGeneticAlgorithmRunner.py
@@ -1,13 +1,9 @@
-from jmetal.algorithm.singleobjective.evolutionaryalgorithm.generationalGeneticAlgorithm import GenerationalGeneticAlgorithm
-from jmetal.core.solution.binarySolution import BinarySolution
-from jmetal.core.solution.floatSolution import FloatSolution
-from jmetal.operator.crossover.SBX import SBX
-from jmetal.operator.crossover.singlepoint import SinglePoint
-from jmetal.operator.mutation.bitflip import BitFlip
-from jmetal.operator.mutation.polynomial import Polynomial
-from jmetal.operator.selection.binarytournament import BinaryTournament
-from jmetal.problem.singleobjective.onemax import OneMax
-from jmetal.problem.singleobjective.sphere import Sphere
+from jmetal.algorithm.singleobjective.evolutionaryalgorithm.evolutionaryalgorithm import GenerationalGeneticAlgorithm
+from jmetal.core.solution import BinarySolution, FloatSolution
+from jmetal.operator.crossover import SinglePoint, SBX
+from jmetal.operator.mutation import BitFlip, Polynomial
+from jmetal.operator.selection import BinaryTournament
+from jmetal.problem.singleobjective import OneMax, Sphere
 
 
 def main():
@@ -22,9 +18,9 @@ def binary_example() -> None:
         problem,
         population_size = 100,
         max_evaluations = 25000,
-        mutation_operator = BitFlip(1.0/bits),
-        crossover_operator = SinglePoint(0.9),
-        selection_operator = BinaryTournament())
+        mutation = BitFlip(1.0/bits),
+        crossover = SinglePoint(0.9),
+        selection = BinaryTournament())
 
     algorithm.run()
     result = algorithm.get_result()
@@ -41,9 +37,9 @@ def run_as_a_thread_example() -> None:
         problem,
         population_size = 100,
         max_evaluations = 25000,
-        mutation_operator = Polynomial(1.0/variables, distribution_index=20),
-        crossover_operator = SBX(1.0, distribution_index=20),
-        selection_operator = BinaryTournament())
+        mutation = Polynomial(1.0/variables, distribution_index=20),
+        crossover = SBX(1.0, distribution_index=20),
+        selection = BinaryTournament())
 
     algorithm.start()
     print("Algorithm (running as a thread): " + algorithm.get_name())
@@ -62,9 +58,9 @@ def float_example() -> None:
         problem,
         population_size = 100,
         max_evaluations = 25000,
-        mutation_operator = Polynomial(1.0/variables, distribution_index=20),
-        crossover_operator = SBX(1.0, distribution_index=20),
-        selection_operator = BinaryTournament())
+        mutation = Polynomial(1.0/variables, distribution_index=20),
+        crossover = SBX(1.0, distribution_index=20),
+        selection = BinaryTournament())
 
     algorithm.run()
     result = algorithm.get_result()
