+++ jmetal/util/observer/test/observerTest.py
@@ -1,7 +1,7 @@
 import unittest
 
-from jmetal.core.util.observer.impl.defaultobservable import DefaultObservable
 from jmetal.core.util.observer.observer import Observer
+from jmetal.util.observer.impl import DefaultObservable
 
 
 class TestMethods(unittest.TestCase):
